mechanism to allow specifing file editing capability
By adding actions="view|edit" to mime/mediatype pixiewood will also add
the EDIT action to the intent handler for that particular mimetype.

This is not specified in the appstream metadata format, but appstreamcli
doesn't check for parameters so it works for now.
I hope to have that behviour codified in appstream at some point.
See https://github.com/ximion/appstream/issues/689