dropped pkg-config 'emulation'
It shouldn't be needed anymore, as meson should now be able to determine
the correct vulkan version for system dependencies even if you can't run
host binaries (ensure you've fetched the latest commit in the android2
branch).
And the latest gtk/android has meson code to manually check for
existance of the needed headerfile.