Support theme generation
The theme is only really relevant for the background of the splash
screen AFAIK, but now they are switchable. Right now "gtk" and "adw" are
supported, but further can by added in generate/themes.json.

The accent color from the theme has been dropped in favour of
/meta:component/meta:branding/meta:color[@type="primary"]. Go use that
instead if you haven't already.