build: Move fingerprint generation to gradle
By letting gradle take care of fingerprint generation, builds with
swapped assets continue to work without having to rerun pixiewood.

This is going to be useful for the interpreted languages runtimes, which
will likely going to be shipped as just the android studio tree, with
users required to swap out their application impl within the assets.

Instead of 128 random bytes like we've done for now, this now computes a
SHA-512 digest of all files (and paths) that are part of the assets and
keeps the other 64 bytes zeroed. This was done to avoid having the app
re-run the copy routine each time it is launched via the run/debug
button in studio.